Move Title and Subtitle to typography
